rm scanner save_to_db; moving fns to db models
