+++ depobs/database/models.py
@@ -1206,3 +1206,174 @@ def get_advisories_by_package_versions(
             [package_version.id for package_version in package_versions]
         )
     )
+
+
+def add_new_package_version(session: sqlalchemy.orm.Session, pkg: Dict) -> None:
+    get_package_version_id_query(session, pkg).one_or_none() or session.add(
+        PackageVersion(
+            name=pkg.get("name", None),
+            version=pkg.get("version", None),
+            language="node",
+            url=pkg.get(
+                "resolved", None
+            ),  # is null for the root for npm list and yarn list output
+        )
+    )
+
+
+def get_package_version_id_query(
+    session: sqlalchemy.orm.Session, pkg: Dict
+) -> sqlalchemy.orm.query.Query:
+    return session.query(PackageVersion.id).filter_by(
+        name=pkg["name"], version=pkg["version"], language="node"
+    )
+
+
+def get_package_version_link_id_query(
+    session: sqlalchemy.orm.Session, link: Tuple[int, int]
+) -> sqlalchemy.orm.query.Query:
+    parent_package_id, child_package_id = link
+    return session.query(PackageLink.id).filter_by(
+        parent_package_id=parent_package_id, child_package_id=child_package_id
+    )
+
+
+def get_node_advisory_id_query(
+    session: sqlalchemy.orm.Session, advisory: Dict
+) -> sqlalchemy.orm.query.Query:
+    return session.query(Advisory.id).filter_by(language="node", url=advisory["url"])
+
+
+def insert_package_graph(session: sqlalchemy.orm.Session, task_data: Dict) -> None:
+    link_ids = []
+    for task_dep in task_data.get("dependencies", []):
+        add_new_package_version(session, task_dep)
+        session.commit()
+        parent_package_id = get_package_version_id_query(session, task_dep).first()
+
+        for dep in task_dep.get("dependencies", []):
+            # is fully qualified semver for npm (or file: or github: url), semver for yarn
+            name, version = dep.rsplit("@", 1)
+            child_package_id = get_package_version_id_query(
+                session, dict(name=name, version=version)
+            ).first()
+
+            link_id = get_package_version_link_id_query(
+                session, (parent_package_id, child_package_id)
+            ).one_or_none()
+            if not link_id:
+                session.add(
+                    PackageLink(
+                        child_package_id=child_package_id,
+                        parent_package_id=parent_package_id,
+                    )
+                )
+                session.commit()
+                link_id = get_package_version_link_id_query(
+                    session, (parent_package_id, child_package_id)
+                ).first()
+            link_ids.append(link_id)
+
+    session.add(
+        PackageGraph(
+            root_package_version_id=get_package_version_id_query(
+                session, task_data["root"]
+            ).first()
+            if task_data["root"]
+            else None,
+            link_ids=link_ids,
+            package_manager="yarn" if "yarn" in task_data["command"] else "npm",
+            package_manager_version=None,
+        )
+    )
+    session.commit()
+
+
+def insert_package_audit(session: sqlalchemy.orm.Session, task_data: Dict) -> None:
+    is_yarn_cmd = bool("yarn" in task_data["command"])
+    # NB: yarn has .advisory and .resolution
+
+    # the same advisory JSON (from the npm DB) is
+    # at .advisories{k, v} for npm and .advisories[].advisory for yarn
+    advisories = (
+        (item.get("advisory", None) for item in task_data.get("advisories", []))
+        if is_yarn_cmd
+        else task_data.get("advisories", dict()).values()
+    )
+    non_null_advisories = (adv for adv in advisories if adv)
+
+    for advisory in non_null_advisories:
+        advisory_fields = extract_nested_fields(
+            advisory,
+            {
+                "package_name": ["module_name"],
+                "npm_advisory_id": ["id"],
+                "vulnerable_versions": ["vulnerable_versions"],
+                "patched_versions": ["patched_versions"],
+                "created": ["created"],
+                "updated": ["updated"],
+                "url": ["url"],
+                "severity": ["severity"],
+                "cves": ["cves"],
+                "cwe": ["cwe"],
+                "exploitability": ["metadata", "exploitability"],
+                "title": ["title"],
+            },
+        )
+        advisory_fields["cwe"] = int(advisory_fields["cwe"].lower().replace("cwe-", ""))
+        advisory_fields["language"] = "node"
+        advisory_fields["vulnerable_package_version_ids"] = []
+
+        get_node_advisory_id_query(
+            session, advisory_fields
+        ).one_or_none() or session.add(Advisory(**advisory_fields))
+        session.commit()
+
+        # TODO: update other advisory fields too
+        impacted_versions = set(
+            finding.get("version", None)
+            for finding in advisory.get("findings", [])
+            if finding.get("version", None)
+        )
+        db_advisory = (
+            session.query(Advisory.id, Advisory.vulnerable_package_version_ids)
+            .filter_by(language="node", url=advisory["url"])
+            .first()
+        )
+        impacted_version_package_ids = list(
+            vid
+            for result in session.query(PackageVersion.id)
+            .filter(
+                PackageVersion.name == advisory_fields["package_name"],
+                PackageVersion.version.in_(impacted_versions),
+            )
+            .all()
+            for vid in result
+        )
+        if len(impacted_versions) != len(impacted_version_package_ids):
+            log.warning(
+                f"missing package versions for {advisory_fields['package_name']!r}"
+                f" in the db or misparsed audit output version:"
+                f" {impacted_versions} {impacted_version_package_ids}"
+            )
+
+        if db_advisory.vulnerable_package_version_ids is None:
+            session.query(Advisory.id).filter_by(id=db_advisory.id).update(
+                dict(vulnerable_package_version_ids=list())
+            )
+
+        # TODO: lock the row?
+        vpvids = set(
+            list(
+                session.query(Advisory)
+                .filter_by(id=db_advisory.id)
+                .first()
+                .vulnerable_package_version_ids
+            )
+        )
+        vpvids.update(set(impacted_version_package_ids))
+
+        session.query(Advisory.id).filter_by(id=db_advisory.id).update(
+            dict(vulnerable_package_version_ids=sorted(vpvids))
+        )
+        session.commit()
