+++ setup.cfg
@@ -34,7 +34,6 @@ files =
       depobs/scanner/pipelines/postprocess.py,
       depobs/scanner/pipelines/run_repo_tasks.py,
       depobs/scanner/pipelines/rust_changelog.py,
-      depobs/scanner/pipelines/save_to_db.py,
       depobs/scanner/pipelines/util.py,
       depobs/database/*.py,
       depobs/util/*.py,
